requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
nltk==3.5
numpy==1.19.1
pandas==1.1.0
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
regex==2020.7.14
requests==2.24.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
tqdm==4.48.2
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
zope.interface==5.1.0
